Make Queryable.execute behave more like Query.execute
The old Queryable.execute implementation has been moved to
Queryable.run. Technically this is an API/ABI break, but I doubt
anyone is using SQLHeavy in a way that will actually break. You would
need to be passing a length to Queryable.execute, and only
Queryable.run_script had a good reason to do that.